+++ src/fields/components/Checkbox/checkbox.props.js
@@ -2,7 +2,13 @@
 import without from 'lodash/without';
 
 // Utils
-import { valuesToOptions } from '../../utils/enum-utils';
+import { UTIL_CONFIG } from '../../utils';
+
+const {
+  ENUM_UTILS: {
+    util: { valuesToOptions, isEnum },
+  },
+} = UTIL_CONFIG;
 
 const doOnChange = (onChange) => (e, checked) => onChange(checked);
 
@@ -19,5 +25,5 @@ export default ({ onChange, schema = {} }) => ({
   onChange: doOnChange(onChange),
   onGroupChange: (value) => onChange && onCheckboxChangeHandler(onChange, value),
   label: schema.title,
-  choices: valuesToOptions(schema.enum),
+  choices: valuesToOptions(isEnum(schema)),
 });