+++ src/fields/components/MaterialSelect/select.props.js
@@ -1,6 +1,14 @@
 // Utils
-import { valuesToOptions } from '../../utils/enum-utils';
-import { coerceValue } from '../../utils/parse-values';
+import { UTIL_CONFIG } from '../../utils';
+
+const {
+  ENUM_UTILS: {
+    util: { valuesToOptions, isEnum },
+  },
+  PARSE_VALUES: {
+    util: { coerceValue, deepStringify },
+  },
+} = UTIL_CONFIG;
 
 const onChangeHandler = (onChange, type) => (e) => {
   const value = coerceValue(type, e.target.value);
@@ -9,7 +17,7 @@ const onChangeHandler = (onChange, type) => (e) => {
 
 export default ({ onChange, schema = {}, type }) => ({
   onChange: onChange && onChangeHandler(onChange, type),
-  choices: valuesToOptions(schema.enum),
+  choices: valuesToOptions(isEnum(schema)),
   nullOption: 'Please select...',
   label: schema.title || '',
 });