+++ src/fields/components/RadioGroup/radio-group.props.js
@@ -1,5 +1,11 @@
 // Utils
-import { valuesToOptions } from '../../utils/enum-utils';
+import { UTIL_CONFIG } from '../../utils';
+
+const {
+  ENUM_UTILS: {
+    util: { valuesToOptions, isEnum },
+  },
+} = UTIL_CONFIG;
 
 const onChangeHandler = (givenOnChange) => (e) => {
   if (e.target.value !== undefined) {
@@ -11,5 +17,5 @@ export default ({ onChange, options = {}, schema = {} }) => ({
   onChange: onChange && onChangeHandler(onChange),
   row: options.inline,
   label: schema.title,
-  choices: valuesToOptions(schema.enum),
+  choices: valuesToOptions(isEnum(schema)),
 });