+++ src/fields/utils/enum-utils/values-to-options.js
@@ -2,8 +2,15 @@ import keys from 'lodash/keys';
 
 export default (values) => {
   if (values instanceof Array) {
+    const returnValueMapping = (value) => {
+      if ('const' in value) {
+        return { key: value.const, value: value.title };
+      }
+      return { ...value };
+    };
     return values.map((e) => {
-      const returnValue = typeof e === 'object' ? { ...e } : { key: e, value: e };
+      const returnValue =
+        typeof e === 'object' ? returnValueMapping(e) : { key: e, value: e };
       return returnValue;
     });
   }